+++ kodosumi/service/app.py
@@ -46,9 +46,15 @@ from kodosumi.service.inputs.timeline.controller import TimelineController
 from kodosumi.service.jwt import JWTAuthenticationMiddleware, operator_guard
 from kodosumi.service.proxy import LockController, ProxyControl
 from kodosumi.service.expose.control import (
-    ExposeControl, ExposeUIControl, BootControl, BootUIControl,
-    MaintenanceControl, ExchangeControl, ExchangeUIControl,
-    AuditLogControl, RegistryControl, WalletsControl, ensure_serve_config
+    ExposeControl, ExposeUIControl, MaintenanceControl, ensure_serve_config
+)
+from kodosumi.service.expose.audit_control import AuditLogControl
+from kodosumi.service.expose.boot_control import BootControl, BootUIControl
+from kodosumi.service.expose.exchange_control import (
+    ExchangeControl, ExchangeUIControl
+)
+from kodosumi.service.expose.registry_control import (
+    RegistryControl, WalletsControl
 )
 from kodosumi.service.expose import db as expose_db
 from kodosumi.service.role import RoleControl, ProfileControl
